achievements: build the response shape once; drop the settings privates
`get_full_state` hand-built the full payload twice — once for the
opted-out shell, once from real state — and the two had already started
to drift in expression: the disabled branch hardcoded `next_threshold`
as `tiers[0]` where the enabled branch computed it. They agreed
numerically, but nothing held them there.

Both branches now go through one pure `_state_payload(state)` renderer.
The opted-out shell is `_state_payload(_ensure_state({}))`: a fresh
empty state fills in zeroed counters and -1 watermarks, which renders as
every category locked, nothing pending, nothing seen — exactly the shell
the old branch spelled out by hand. Two tests pin the invariant, one of
which asserts the opted-out payload is identical to a zero-progress
user's.

The module also reached into three `vtsearch.settings` privates
(`_ensure_user_loaded`, `_settings_lock`, and a function-local
`_user_caches` import) to take a consistent read snapshot. That pattern
is now a public `settings.snapshot_user(username)`, which owns the
lock-ordering rule that made the open-coded version delicate;
`get_user_settings` uses it too. No production module reads those
containers any more — only the tests that assert on them — so the
comment in settings.py explaining why they stay module-global is
narrowed to say so.

Finally, the static catalog (categories, tiers, media types, hours,
Readme Reader docs) moves to `vtsearch/achievements_catalog.py`, leaving
`achievements.py` as the state machine. Every public name is re-exported
via `__all__`, so callers keep importing them from `vtsearch.achievements`.

Closes #3430
